add simple gnumeric sheet for calculation
You can try out how modifications of coefficients, parameters and values
influence the result. The formula is taken from the data sheet (without the
error). There may be a difference of a few digits compared to the C
implementation as a result of internal rounding as gnumeric calculates with
floating point numbers.